rename init to newPage to be less confusing
