+++ libs/scraper/src/puppeteer-init/index.ts
@@ -1,4 +1,4 @@
-export * from './init';
+export * from './newPage';
 export * from './rootFolder';
 export * from './userProfile';
 export * from './browser';
