+++ libs/scraper/src/record/registry.ts
@@ -1,5 +1,5 @@
 import type { Browser } from 'puppeteer';
-import { newPage } from '../puppeteer-init/init';
+import { newPage } from '../puppeteer-init/newPage';
 import { log } from '@thecointech/logging';
 import { RecorderOptions } from './types';
 import { Recorder } from './recorder';
